debundle/lowering: extract naturalize
BAZEL_TEST_INVOCATIONS=none: file move.
